+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - `fullSize`
   - `fullSizeOnMobile`
 - Feature: `Modal/Next`: Deprecate `withoutMargin` prop on `Modal.Paragraph` in favor of `noMargin`.
+- Feature: Add `boron` modifier to Button component.
 - Feature: Add new color: `font3`.
 
 ## [3.11.0] - 2021-05-14